requirements.txt
mysql-connector-python~=8.0.31
bcrypt~=4.0.1
pyserial~=3.5
guizero~=1.3.0